New namespace, PCA9554 driver update, INA219 driver update
The new namespace in INTERRUPT lib, name is INTERRUPT. This is needed for example to add new interrupts for other chips.
PCA9554:  fix bug and optimization code. Update save i2c bus number.
INA219: add new function - change address. This is needed for example if you have some INA chips on one I2C bus and so as not to create a new object for each.